Their wave split `tests/test_code_audit.py` into `tests/code_audit/*`, so the two
new tests moved into `tests/code_audit/test_default_via_or.py` beside the rest of
that scanner's suite rather than resurrecting the deleted file. The baseline
conflict is resolved by taking their reworked version as the base and re-applying
only the two line-shift re-keys this change causes.

Verified semantically rather than by the absence of conflict markers, since
`default_via_or.py` auto-merged after a 159-line rewrite on their side: the file
parses, `_site`/`_enclosing_symbols`/`_classify` each survive exactly once, and
neither the generic detail literal nor the placeholder swallow came back.

Three failures on origin/master predate this merge and block any commit there, so
they are fixed here rather than worked around: the mypy source-file count in
`.pre-commit-config.yaml` and `mypy-full.yml` (their merge moved it back to 248),
and `_catalogue.py`'s `raw in (None, "")` guard, which does not narrow for mypy -
`raw is None or raw == ""` does, and is the same test.

+++ src/pyutilz/data/polarslib/binning.py
@@ -57,7 +57,7 @@ def _group_freqs(bins: pl.DataFrame, cols, drop_nulls: bool = False) -> Any:
 
 def _shannon_entropy(freqs: Any) -> float:
     """Shannon entropy (in nats) of a discrete frequency distribution that sums to 1."""
-    return -np.sum(freqs * np.log(freqs))  # type: ignore[no-any-return]  # untyped upstream source (json/external lib/dynamic attr); return value verified correct at runtime
+    return float(-np.sum(freqs * np.log(freqs)))  # float(): np.sum over an Any operand returns np.floating, not the declared float
 
 
 def entropy_for_column(bins: pl.DataFrame, col: str, drop_nulls: bool = False) -> float:
@@ -267,7 +267,9 @@ def bin_numerical_columns(
         skipped_clips = []
         if min_nuniques_to_clip:
             # do not apply clipping if # of unique values is too low (under 10)
-            n_uniques_dict = df.lazy().select(pl.col(clips.keys()).n_unique()).collect().row(0, named=True)
+            # Cast before counting: older polars refuses n_unique on Decimal, and the count is
+            # identical either way because the cast is value-preserving for the widths used here.
+            n_uniques_dict = df.lazy().select(pl.col(clips.keys()).cast(pl.Float64).n_unique()).collect().row(0, named=True)
             for col, nuniques in n_uniques_dict.items():
                 if nuniques < min_nuniques_to_clip:
                     for field in "min max".split():
